Create custom-css
First commit for custom css. The original css is styles.css, part of the Serene Wordpress Theme by Elegant Themes. Minor edits to the header. Most relevant here is the "clippable" class. Your pull requests are welcome for the #paella id name, whose purpose is to show no bullets in a <ul> bullet list; I initially could not create a class because it was overridden by a native class called "article.post ul".